+++ libs/app/cjs/ServiceContainer.js
@@ -191,7 +191,8 @@ const configOverrider = (defConf, envConf)=>{
      * @param {array} args - Array of path parts
      * @returns {string}
      */ toAbsolutePath(...args) {
-        if (args.length === 0 || args[0] == null) {
+        args = args.filter((arg)=>arg != null);
+        if (args.length === 0) {
             return this.workingPath;
         }
         return _nodepath.default.resolve(this.workingPath, ...args);
@@ -368,7 +369,7 @@ const configOverrider = (defConf, envConf)=>{
         // if no logger in config, use console logger
         this.flushLogCache();
         await this.emit_('configFinalized', this.config);
-        if (this.options.logLevel === 'debug' || this.options.logLevel === 'verbose') {
+        if (this.options.logConfig && (this.options.logLevel === 'debug' || this.options.logLevel === 'verbose')) {
             this.log('verbose', 'Finalized config:', this.config);
         }
         let featureGroups = {
@@ -510,8 +511,12 @@ const configOverrider = (defConf, envConf)=>{
          * @member {string}
          */ this.configPath = this.toAbsolutePath(this.options.configPath);
         /**
+         * Source files path.
+         * @member {string}
+         **/ this.sourcePath = this.toAbsolutePath(this.options.sourcePath);
+        /**
          * Feature path
-         */ this.featuresPath = this.toAbsolutePath(this.options.featuresPath);
+         */ this.featuresPath = _nodepath.default.resolve(this.sourcePath, this.options.featuresPath);
         this._logCache = [];
         // dummy
         this.log = (...args)=>{