+++ libs/app/lib/ModuleBase.js
@@ -29,12 +29,6 @@ const ModuleBase = (Base) =>
              **/
             this.isServer = false;
 
-            /**
-             * Source files path.
-             * @member {string}
-             **/
-            this.sourcePath = this.toAbsolutePath(this.options.sourcePath);
-
             this.featuresPath = path.resolve(this.sourcePath, this.options.featuresPath);
 
             this.logger = this.host.logger?.child({ module: this.name }, { level: this.options.logLevel });