+++ libs/types/test/issues.spec.js
@@ -1,0 +1,7 @@
+import { Types } from "..";
+
+describe('issues', () => {
+    it('sanitize without meta failed', () => {
+        Types.INTEGER.sanitize('123').should.equal(123);
+    });
+});